adhd-api/requirements.txt
# 基础依赖
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0

# 数据库
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# 认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# 数据验证
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# CORS
python-multipart==0.0.6
